Breakup components
