+++ frontend/app/pokemon/[id]/components/PokemonAbility/index.tsx
@@ -1,0 +1,1 @@
+export { PokemonAbility as default } from "./PokemonAbility";
