+++ frontend/app/pokemon/[id]/components/PokemonDetail/PokemonDetail.module.css
@@ -171,110 +171,6 @@
   margin-bottom: 1rem;
 }
 
-.typeBadge {
-  padding: 0.375rem 0.625rem;
-  font-size: 0.75rem;
-  font-weight: 600;
-  border-radius: 0.75rem;
-  text-transform: uppercase;
-  letter-spacing: 0.25px;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-  min-width: 55px;
-  text-align: center;
-}
-
-/* Type-specific colors - matching Pokemon card styles */
-.typeBadge.normal {
-  background: #a8a878;
-  color: #fff;
-}
-
-.typeBadge.fire {
-  background: #f08030;
-  color: #fff;
-}
-
-.typeBadge.water {
-  background: #6890f0;
-  color: #fff;
-}
-
-.typeBadge.electric {
-  background: #f8d030;
-  color: #000;
-}
-
-.typeBadge.grass {
-  background: #78c850;
-  color: #fff;
-}
-
-.typeBadge.ice {
-  background: #98d8d8;
-  color: #000;
-}
-
-.typeBadge.fighting {
-  background: #c03028;
-  color: #fff;
-}
-
-.typeBadge.poison {
-  background: #a040a0;
-  color: #fff;
-}
-
-.typeBadge.ground {
-  background: #e0c068;
-  color: #000;
-}
-
-.typeBadge.flying {
-  background: #a890f0;
-  color: #fff;
-}
-
-.typeBadge.psychic {
-  background: #f85888;
-  color: #fff;
-}
-
-.typeBadge.bug {
-  background: #a8b820;
-  color: #fff;
-}
-
-.typeBadge.rock {
-  background: #b8a038;
-  color: #fff;
-}
-
-.typeBadge.ghost {
-  background: #705898;
-  color: #fff;
-}
-
-.typeBadge.dragon {
-  background: #7038f8;
-  color: #fff;
-}
-
-.typeBadge.dark {
-  background: #705848;
-  color: #fff;
-}
-
-.typeBadge.steel {
-  background: #b8b8d0;
-  color: #000;
-}
-
-.typeBadge.fairy {
-  background: #ee99ac;
-  color: #fff;
-}
-
 .statsSection {
   background: rgba(255, 255, 255, 0.9);
   padding: 1.5rem;
@@ -287,46 +183,6 @@
   color: #212529;
   font-size: 1.25rem;
   font-weight: 700;
-}
-
-.statsGrid {
-  display: grid;
-  gap: 1rem;
-}
-
-.statItem {
-  display: grid;
-  grid-template-columns: 80px 1fr 50px;
-  align-items: center;
-  gap: 15px;
-}
-
-.statLabel {
-  font-weight: 600;
-  color: #495057;
-  font-size: 0.875rem;
-}
-
-.statBar {
-  height: 12px;
-  background: #e9ecef;
-  border-radius: 6px;
-  overflow: hidden;
-  position: relative;
-}
-
-.statFill {
-  height: 100%;
-  background: #0d6efd;
-  border-radius: 6px;
-  transition: width 0.8s ease;
-}
-
-.statValue {
-  font-weight: 700;
-  color: #212529;
-  text-align: right;
-  font-size: 0.875rem;
 }
 
 .abilitiesSection {
@@ -360,70 +216,6 @@
 .abilitiesGrid {
   display: grid;
   gap: 1rem;
-}
-
-.abilityCard {
-  background: var(--card-bg, #fff);
-  border: 1px solid #dee2e6;
-  border-radius: 0.5rem;
-  padding: 1rem;
-  transition: all 0.2s ease;
-}
-
-.abilityCard:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
-}
-
-.abilityHeader {
-  display: flex;
-  align-items: center;
-  gap: 15px;
-  margin-bottom: 15px;
-  flex-wrap: wrap;
-}
-
-.abilityName {
-  margin: 0;
-  color: #212529;
-  font-size: 1.125rem;
-  font-weight: 600;
-  text-transform: capitalize;
-  letter-spacing: 0.5px;
-}
-
-.abilitySlot {
-  background: #6c757d;
-  color: white;
-  padding: 0.25rem 0.75rem;
-  border-radius: 0.75rem;
-  font-size: 0.75rem;
-  font-weight: 600;
-}
-
-.abilityDetails {
-  border-top: 1px solid #dee2e6;
-  padding-top: 15px;
-}
-
-.abilityDescription {
-  margin: 0 0 10px 0;
-  color: #495057;
-  line-height: 1.6;
-  font-size: 0.875rem;
-}
-
-.abilityEffect,
-.abilityGeneration {
-  margin: 5px 0;
-  color: #6c757d;
-  font-size: 0.875rem;
-  line-height: 1.5;
-}
-
-.abilityEffect strong,
-.abilityGeneration strong {
-  color: #495057;
 }
 
 /* Responsive design */
