+++ frontend/app/pokemon/[id]/components/PokemonDetail/PokemonDetailSkeleton.module.css
