+++ frontend/app/pokemon/[id]/components/PokemonDetail/PokemonDetailSkeleton.tsx
