+++ frontend/app/pokemon/[id]/components/PokemonDetail/index.ts
