+++ frontend/app/pokemon/[id]/components/PokemonStat/index.ts
@@ -1,0 +1,1 @@
+export { PokemonStat as default } from "./PokemonStat";
