+++ frontend/app/ui/PokemonCard/PokemonCard.module.css
@@ -146,110 +146,6 @@
   justify-content: center;
 }
 
-.type {
-  padding: 0.375rem 0.625rem;
-  font-size: 0.75rem;
-  font-weight: 600;
-  border-radius: 0.75rem;
-  text-transform: uppercase;
-  letter-spacing: 0.25px;
-  border: 1px solid rgba(0, 0, 0, 0.1);
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
-  min-width: 55px;
-  text-align: center;
-}
-
-/* Type-specific colors - more muted and authentic */
-.type.normal {
-  background: #a8a878;
-  color: #fff;
-}
-
-.type.fire {
-  background: #f08030;
-  color: #fff;
-}
-
-.type.water {
-  background: #6890f0;
-  color: #fff;
-}
-
-.type.electric {
-  background: #f8d030;
-  color: #000;
-}
-
-.type.grass {
-  background: #78c850;
-  color: #fff;
-}
-
-.type.ice {
-  background: #98d8d8;
-  color: #000;
-}
-
-.type.fighting {
-  background: #c03028;
-  color: #fff;
-}
-
-.type.poison {
-  background: #a040a0;
-  color: #fff;
-}
-
-.type.ground {
-  background: #e0c068;
-  color: #000;
-}
-
-.type.flying {
-  background: #a890f0;
-  color: #fff;
-}
-
-.type.psychic {
-  background: #f85888;
-  color: #fff;
-}
-
-.type.bug {
-  background: #a8b820;
-  color: #fff;
-}
-
-.type.rock {
-  background: #b8a038;
-  color: #fff;
-}
-
-.type.ghost {
-  background: #705898;
-  color: #fff;
-}
-
-.type.dragon {
-  background: #7038f8;
-  color: #fff;
-}
-
-.type.dark {
-  background: #705848;
-  color: #fff;
-}
-
-.type.steel {
-  background: #b8b8d0;
-  color: #000;
-}
-
-.type.fairy {
-  background: #ee99ac;
-  color: #fff;
-}
-
 .stats {
   font-size: 0.875rem;
   color: #495057;
