+++ frontend/app/ui/PokemonTypePill/PokemonTypePill.tsx
@@ -1,0 +1,16 @@
+import styles from "./PokemonTypePill.module.css";
+
+interface PokemonTypePillProps {
+  type: string;
+}
+
+export const PokemonTypePill = ({ type }: PokemonTypePillProps) => {
+  return (
+    <span
+      key={type}
+      className={`${styles.typePill} ${styles[type.toLowerCase()]}`}
+    >
+      {type}
+    </span>
+  );
+};
