+++ frontend/app/ui/PokemonTypePill/index.tsx
@@ -1,0 +1,1 @@
+export { PokemonTypePill as default } from "./PokemonTypePill";
